210204 pull request

+++ src/main/java/org/zerock/domain/CustomerVO.java
@@ -4,7 +4,7 @@ import lombok.Data;
 
 @Data
 public class CustomerVO {
-	private int customerCode;
+	private long customerCode;
 	private String customerEmail;
 	private String customerAddress;
 	private String customerPhone;
